refactor: break out more files
